requirements.txt
django==3.2.9
djangorestframework==3.12.4
django-environ==0.8.1
drf-yasg==1.20.0
